sqldf-package {sqldf}	R Documentation

sqldf package overview

Description

Provides an easy way to perform SQL selects on R data frames.

Details

The package contains a single function sqldf whose help file contains more information and examples.

References

The sqldf help page contains the primary documentation. The sqldf home page http://sqldf.googlecode.com contains links to SQLite pages that may be helpful in formulating queries.

[Package sqldf version 0.4-10 Index]